backend/requirements.txt
# Django 4.2+ LTS
django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.4.0
django-filter==24.3
django-environ==0.12.0
django-extensions==3.2.3

# PostgreSQL driver
psycopg[binary]==3.2.4
psycopg2-binary==2.9.10

# Production server
gunicorn==23.0.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-env==1.1.5
factory-boy==3.3.1

# Code quality
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

